* created activity_module for the activity stream
