+++ public/modules/activity/AbstractActivity.php
@@ -1,0 +1,5 @@
+<?php
+
+abstract class AbstractActivity {
+    
+}
