+++ public/modules/autoload.php
@@ -8,14 +8,15 @@ require_once('profile/Profile_Characters.php');
 require_once('settings/Settings.php');
 require_once('identity/Identity.php');
 require_once('activity/Activity.php');
-require_once('activity/Comment.php');
 require_once('activity/shout/Activity_Shout.php');
 require_once('activity/event/Activity_Event.php');
+require_once('activity/Activity_Stream.php');
 require_once('activity/event/widgets/Activity_Event_Widgets.php');
 require_once('activity/event/signups/Activity_Event_Signups.php');
 require_once('activity/event/signups/widgets/Activity_Event_Signups_Widgets.php');
 require_once('activity/poll/Activity_Poll.php');
 require_once('activity/actionmsg/Activity_ActionMsg.php');
+require_once('activity/Comment.php');
 require_once('pages/Pages.php');
 require_once('gw2api/gw2api.php');
 require_once('gw2api/widgets/gw2api_Widgets.php');
